backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==3.13.0
django-filter==2.4.0
django-mixins==0.0.10
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
webcolors==1.13
